text_to_image/requirements.txt
pandas==0.25.3
scikit-learn==0.21.3
Pillow==6.2.1
requests==2.22.0
opencv-python==4.1.2.30